calculaterank実装、コントローラーとの連携

+++ app/Service/CalculateRankInterface.php
@@ -5,5 +5,5 @@ namespace App\Service;
 interface CalculateRankInterface
 {
     public function evaluation(string $name): array;
-    public function normalcdf(int $mean, int $sigma, int $to): int;
+    public function normalcdf(int $mean, int $sigma, int $to): float;
 }
